+++ src/app/(chat)/layout.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import AppFrame from "@app/_component/AppFrame";
+import Menu from "@app/_component/Menu";
 import WithAuth from "@app/_hoc/WithAuth";
-import Menu from "@components/sider/Menu";
 
 function ChatLayout({ user, children }) {
     return <AppFrame menu={<Menu user={user} />}>{children}</AppFrame>;
