+++ src/app/page.tsx
@@ -1,11 +1,11 @@
 "use client";
 
 import AppFrame from "@app/_component/AppFrame";
+import Menu from "@app/_component/Menu";
 import WithAuth from "@app/_hoc/WithAuth";
 import ChatFrame from "@app/(chat)/rooms/[roomId]/_component/ChatFrame";
 import Dialog from "@app/(chat)/rooms/[roomId]/_component/Dialog";
 import Textator from "@app/(chat)/rooms/[roomId]/_component/Textator";
-import Menu from "@components/sider/Menu";
 
 function Home({ user }) {
     return (
