--- src/components/layout/ContentFrame.tsx
@@ -1,3 +1,0 @@
-export default function ContentFrame() {
-    return <div>ContentFrame</div>;
-}
